fix(input): input lenght can now be over INT_MAX int is not used anymore as a string lenght or as an index. size_t is used instead. getstdin variable u now uses SIZE_MAX as a non-value instead of -1. also, removed some irrelevant instructions in header files

+++ include/display.h
@@ -6,14 +6,15 @@
 /*   By: bprunevi <bprunevi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/03 15:26:07 by bprunevi          #+#    #+#             */
-/*   Updated: 2019/10/15 16:44:33 by bprunevi         ###   ########.fr       */
+/*   Updated: 2019/10/20 11:26:11 by bprunevi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef DISPLAY_H
 #define DISPLAY_H
+#include "libft.h"
 
 void display_init(int sig);
-int display(char *str, int j, int i, int u, char *prompt, int prompt_len);
+int display(char *str, size_t j, size_t i, size_t u, char *prompt, size_t prompt_len);
 
 #endif
